requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
opencv-python==4.8.1.78
Pillow==10.0.1
tqdm==4.66.1
python-dotenv==1.0.0

# PyTorch with CUDA support (install with: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118)
# For CPU-only: pip install torch torchvision torchaudio
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# Machine Learning & AI
transformers==4.35.2
ultralytics==8.0.196
easyocr==1.7.0
fasttext==0.9.2

# CLIP (fix import issues)
git+https://github.com/openai/CLIP.git

# API clients
apify-client==1.7.1
google-generativeai==0.3.2

# Audio processing
openai-whisper==20231117

# Web scraping & utilities
requests==2.31.0
beautifulsoup4==4.12.2
yt-dlp==2023.11.16

# Data processing
scikit-learn==1.3.2
scipy==1.11.4

# Video processing
mediapipe==0.10.7

# Additional dependencies for enhanced features
matplotlib==3.7.2
seaborn==0.12.2

# Custom modules (create these if missing)
# mine_redis - implement your video download logic
# aesthetic_predictor - implement or use fallback